feat!: bump version as previous commit wasn't detected
BREAKING CHANGE: major project refactor

This pull request has grown larger then originally anticipated and should have been multiple pull requests, however this refactors the entire project structure, implements version, blueprint extraction and compilation for the explorer.

The explorer will assists in development by making it easy to track down entities and piece the structure of the planner together, it is not a deployable asset due to the blueprints coming in at 180MB. The hosting of this would not be feasible and is not required, this will simply be served locally where the file downloads in a matter of milliseconds as it's running off the file system.

If users wish to explore the blueprints then the project can be cloned, built and run locally themself.